Improved the performances for the comment component

+++ Solutions/Business.StarterIntranet/app/src/components/Comments/DiscussionBoard/DiscussionBoard.tsx
@@ -6,13 +6,16 @@ import IDiscussion from "../../../models/IDiscussion";
 import { Web, PermissionKind } from "sp-pnp-js";
 import DiscussionReply from "../DiscussionReply/DiscussionReply";
 import { IDiscussionReply, DiscussionPermissionLevel } from "../../../models/IDiscussionReply";
+import * as immutability from "immutability-helper";
+
+// Needed to get it work at runtime
+const update = immutability as any;
 
 class DiscussionBoard extends React.Component<IDiscussionBoardProps, IDiscussionBoardState> {
 
     private _socialModule: SocialModule;
     private _associatedPageId: number;
     private _parentId: number;
-    private _commentInputRef: any;
     private _dicussionBoardListRelativeUrl: string;
 
     public constructor() {
@@ -22,6 +25,7 @@ class DiscussionBoard extends React.Component<IDiscussionBoardProps, IDiscussion
         this.state = {
             discussion: null,
             userPermissions: [],
+            inputValue: ""
         };
 
         this._dicussionBoardListRelativeUrl = `${_spPageContextInfo.webServerRelativeUrl}/Lists/Comments`;
@@ -31,6 +35,8 @@ class DiscussionBoard extends React.Component<IDiscussionBoardProps, IDiscussion
         this.addNewComment = this.addNewComment.bind(this);
         this.deleteReply = this.deleteReply.bind(this);
         this.updateReply = this.updateReply.bind(this);
+
+        this.onValueChange = this.onValueChange.bind(this);
     }
 
     public render() {
@@ -48,17 +54,16 @@ class DiscussionBoard extends React.Component<IDiscussionBoardProps, IDiscussion
         // If the current user can add list item to the list, it means he can comment
         if (this.state.userPermissions.indexOf(DiscussionPermissionLevel.Add) !== -1) {
             renderNewReply = <div>
-                <textarea ref={ (input) => {
-                    this._commentInputRef = input;
-                }}></textarea>
+                <textarea value={ this.state.inputValue } onChange={ this.onValueChange } placeholder="Add your comment..."></textarea>
                 <button type="button" onClick={ () => { 
-
+       
                     let parentId = null;
                     if (this.state.discussion) {
                         parentId = this.state.discussion.Id;
                     }
 
-                    this.addNewComment(parentId, this._commentInputRef.value) 
+                    this.addNewComment(parentId, this.state.inputValue);
+                    
                 }}>Add new comment</button>
             </div>
         }
@@ -67,6 +72,10 @@ class DiscussionBoard extends React.Component<IDiscussionBoardProps, IDiscussion
             { renderPageComments }
             { renderNewReply }
         </div>
+    }
+
+    public onValueChange(e: any) {
+        this.setState({ inputValue: e.target.value });
     }
 
     public async componentDidMount() {
@@ -89,43 +98,70 @@ class DiscussionBoard extends React.Component<IDiscussionBoardProps, IDiscussion
 
     public async addNewComment(parentId: number, replyBody: string) {
 
-        // First comment will create a new discussion
-        if (!parentId) {
-            parentId = await this.createNewDiscussion($("#title").text(), window.location.href);
+        if (!this.state.inputValue) {
+            alert("You can't post an empty comment");
+        } else {
+            let currentDiscussion = this.state.discussion;
+
+            // First comment will create a new discussion
+            if (!parentId) {
+                const newDiscussion = await this.createNewDiscussion($("#title").text(), window.location.href);
+                currentDiscussion = update(currentDiscussion, { $set: newDiscussion});
+
+                // Set the new parent Id
+                parentId = newDiscussion.Id;
+            }
+
+            // Create reply to the discussion and and it to the state
+            // Set the content as HTML (default field type)
+            const reply = await this.createNewDiscussionReply(parentId, `<div>${replyBody}</div>`);
+            currentDiscussion = update(currentDiscussion, { Replies: { $push: [reply]} });
+
+            // Update the discussion
+            this.setState({
+                discussion: currentDiscussion,
+                inputValue: ""
+            });
         }
-
-        // Create reply to the discussion
-        await this.createNewDiscussionReply(parentId, replyBody);
-
-        // Fetch the updated discussion
-        await this.getPageDiscussion(this._associatedPageId);
     }
 
     public async deleteReply(replyId: number) {
-        await this._socialModule.deleteReply(replyId);
 
-        // Fetch the updated discussion
-        await this.getPageDiscussion(this._associatedPageId);
+        if (confirm('Are you sure you want to delete this comment?')) {
+            await this._socialModule.deleteReply(replyId);
+
+            const updatedReplies = this.state.discussion.Replies.filter((reply) => {
+                return reply.Id !== replyId;
+            });
+
+            // Update state
+            this.setState({
+                discussion: update(this.state.discussion, { Replies: { $set: updatedReplies }}),
+            });
+        } 
     }
 
-    public async updateReply(reply: IDiscussionReply) {
-        await this._socialModule.updateReply(reply);
+    public async updateReply(replyToUpdate: IDiscussionReply) {
 
-        // Fetch the updated discussion
-        await this.getPageDiscussion(this._associatedPageId);
+        await this._socialModule.updateReply(replyToUpdate.Id, replyToUpdate.Body);
+
+        const updatedReplies = this.state.discussion.Replies.map((currentReply) => {
+
+            let updatedReply = currentReply;
+            if (currentReply.Id === replyToUpdate.Id) {
+                updatedReply.Body = replyToUpdate.Body;
+            }
+            return updatedReply;
+        });
+
+        // Update state
+        this.setState({
+            discussion: update(this.state.discussion, { Replies: { $set: updatedReplies }}),
+        });
     }
 
-    private async createNewDiscussion(title: string, body: string): Promise<number>{
-
-        const newDiscussionInfo: IDiscussion = {
-            AssociatedPageId: this._associatedPageId,
-            Body: body,
-            Title: title,
-        };
-
-        const discussionId = await this._socialModule.createNewDiscussion(newDiscussionInfo);
-
-        return discussionId;
+    private async createNewDiscussion(title: string, body: string): Promise<IDiscussion>{
+        return await this._socialModule.createNewDiscussion(this._associatedPageId, title, body);
     }
 
     private async getPageDiscussion(associatedPageId: number) {
@@ -137,7 +173,7 @@ class DiscussionBoard extends React.Component<IDiscussionBoardProps, IDiscussion
         });
     }
 
-    private async createNewDiscussionReply(parentId: number, replyBody: string): Promise<void> {
+    private async createNewDiscussionReply(parentId: number, replyBody: string): Promise<IDiscussionReply> {
         return await this._socialModule.createNewDiscussionReply(parentId, replyBody);
     }
